Toolchain: update to support QNN SDK with lib-safe and FastRPC BCS
1. Update to support QNN SDK with lib-safe.
2. QNX toolchain add HAP related header files for FastRPC BCS.

+++ scripts/build/build-target.sh
@@ -60,6 +60,10 @@ if ! [[ -v ENABLE_CTC ]] ; then
   export CTC_BUILD_PREFIX=""
 fi
 
+if ! [[ -v ENABLE_SIMNODE ]] ; then
+  export ENABLE_SIMNODE=OFF
+fi
+
 if ! [[ -v ENABLE_C2D ]] ; then
   if [[ "${QC_TARGET_SOC}" == "8797" ]] ; then
     export ENABLE_C2D=OFF
@@ -373,7 +377,11 @@ setup_qnn_sdk() {
     if [[ -f ${QNN_SDK_ROOT}/lib/${qnx_target_arch}/libGenie.so ]] ; then
       mkdir -p $destdir/opt/qcnode/include
       mkdir -p $destdir/opt/qcnode/lib
-      cp -v ${QNN_SDK_ROOT}/lib/${qnx_target_arch}/libGenie.so $destdir/opt/qcnode/lib
+      if [ -f ${QNN_SDK_ROOT}/lib-safe/${qnx_target_arch}/libGenie.so ]; then
+        cp -v ${QNN_SDK_ROOT}/lib-safe/${qnx_target_arch}/libGenie.so $destdir/opt/qcnode/lib
+      else
+        cp -v ${QNN_SDK_ROOT}/lib/${qnx_target_arch}/libGenie.so $destdir/opt/qcnode/lib
+      fi
       cp -v ${QNN_SDK_ROOT}/include/Genie/*.h $destdir/opt/qcnode/include
       export ENABLE_GENIE=ON
     fi
@@ -455,6 +463,7 @@ cmake \
     -DCMAKE_EXPORT_COMPILE_COMMANDS=ON \
     -DENABLE_GCOV=${ENABLE_GCOV} \
     -DENABLE_CTC=${ENABLE_CTC} \
+    -DENABLE_SIMNODE=${ENABLE_SIMNODE} \
     -DENABLE_C2D=${ENABLE_C2D} \
     -DENABLE_EVA=${ENABLE_EVA} \
     -DENABLE_EVA_AUTO=${ENABLE_EVA_AUTO} \
@@ -492,18 +501,42 @@ aarch64-qnx)
       QNX_VARIANT=aarch64-qnx
     fi
     cp -vf $QNN_SDK_ROOT/bin/${QNX_VARIANT}/* $destdir/opt/qcnode/bin
-    cp -vf $QNN_SDK_ROOT/lib/${QNX_VARIANT}/libQnn* $destdir/opt/qcnode/lib
-    cp -vf $QNN_SDK_ROOT/lib/${HEXAGON_VARIANT}/unsigned/libQnn* $destdir/opt/qcnode/lib/dsp
+    if [ -d $QNN_SDK_ROOT/lib-safe/${QNX_VARIANT} ]; then
+      cp -v $QNN_SDK_ROOT/lib-safe/${QNX_VARIANT}/libQnn* $destdir/opt/qcnode/lib
+    else
+      cp -vf $QNN_SDK_ROOT/lib/${QNX_VARIANT}/libQnn* $destdir/opt/qcnode/lib
+    fi
+    if [ -d $QNN_SDK_ROOT/lib-safe/${HEXAGON_VARIANT} ]; then
+      cp -vf $QNN_SDK_ROOT/lib-safe/${HEXAGON_VARIANT}/unsigned/libQnn* $destdir/opt/qcnode/lib/dsp
+    else
+      cp -vf $QNN_SDK_ROOT/lib/${HEXAGON_VARIANT}/unsigned/libQnn* $destdir/opt/qcnode/lib/dsp
+    fi
     ;;
 aarch64-linux)
     cp -vf $QNN_SDK_ROOT/bin/aarch64-${variant}-linux-gcc11.2/* $destdir/opt/qcnode/bin
-    cp -vf $QNN_SDK_ROOT/lib/aarch64-${variant}-linux-gcc11.2/libQnn* $destdir/opt/qcnode/lib
-    cp -vf $QNN_SDK_ROOT/lib/${HEXAGON_VARIANT}/unsigned/libQnn* $destdir/opt/qcnode/lib/dsp
+    if [ -d $QNN_SDK_ROOT/lib-safe/aarch64-${variant}-linux-gcc11.2 ]; then
+      cp -vf $QNN_SDK_ROOT/lib-safe/aarch64-${variant}-linux-gcc11.2/libQnn* $destdir/opt/qcnode/lib
+    else
+      cp -vf $QNN_SDK_ROOT/lib/aarch64-${variant}-linux-gcc11.2/libQnn* $destdir/opt/qcnode/lib
+    fi
+    if [ -d $QNN_SDK_ROOT/lib-safe/${HEXAGON_VARIANT} ]; then
+      cp -vf $QNN_SDK_ROOT/lib-safe/${HEXAGON_VARIANT}/unsigned/libQnn* $destdir/opt/qcnode/lib/dsp
+    else
+      cp -vf $QNN_SDK_ROOT/lib/${HEXAGON_VARIANT}/unsigned/libQnn* $destdir/opt/qcnode/lib/dsp
+    fi
     ;;
 aarch64-ubuntu)
     cp -vf $QNN_SDK_ROOT/bin/aarch64-${variant}-linux-gcc11.2/* $destdir/opt/qcnode/bin
-    cp -vf $QNN_SDK_ROOT/lib/aarch64-${variant}-linux-gcc11.2/libQnn* $destdir/opt/qcnode/lib
-    cp -vf $QNN_SDK_ROOT/lib/${HEXAGON_VARIANT}/unsigned/libQnn* $destdir/opt/qcnode/lib/dsp
+    if [ -d $QNN_SDK_ROOT/lib-safe/aarch64-${variant}-linux-gcc11.2 ]; then
+      cp -vf $QNN_SDK_ROOT/lib-safe/aarch64-${variant}-linux-gcc11.2/libQnn* $destdir/opt/qcnode/lib
+    else
+      cp -vf $QNN_SDK_ROOT/lib/aarch64-${variant}-linux-gcc11.2/libQnn* $destdir/opt/qcnode/lib
+    fi
+    if [ -d $QNN_SDK_ROOT/lib-safe/${HEXAGON_VARIANT} ]; then
+      cp -vf $QNN_SDK_ROOT/lib-safe/${HEXAGON_VARIANT}/unsigned/libQnn* $destdir/opt/qcnode/lib/dsp
+    else
+      cp -vf $QNN_SDK_ROOT/lib/${HEXAGON_VARIANT}/unsigned/libQnn* $destdir/opt/qcnode/lib/dsp
+    fi
     ;;
 esac
 fi
